Fix GP32 build
darcs-hash:20050427183658-10d61-d1a8b1fd115bbeb0943f7fb9b6f649b86d19eb80.gz

+++ wd2797.h
@@ -11,16 +11,16 @@
 void wd2797_init(void);
 void wd2797_reset(void);
 int wd2797_load_disk(char *filename, int drive);
-void wd2797_command_write(uint_fast8_t octet);
-void wd2797_track_register_write(uint_fast8_t octet);
-void wd2797_sector_register_write(uint_fast8_t octet);
-void wd2797_data_register_write(uint_fast8_t octet);
-uint_fast8_t wd2797_status_read(void);
-uint_fast8_t wd2797_track_register_read(void);
-uint_fast8_t wd2797_sector_register_read(void);
-uint_fast8_t wd2797_data_register_read(void);
+void wd2797_command_write(unsigned int octet);
+void wd2797_track_register_write(unsigned int octet);
+void wd2797_sector_register_write(unsigned int octet);
+void wd2797_data_register_write(unsigned int octet);
+unsigned int wd2797_status_read(void);
+unsigned int wd2797_track_register_read(void);
+unsigned int wd2797_sector_register_read(void);
+unsigned int wd2797_data_register_read(void);
 void wd2797_generate_interrupt(void);
-void wd2797_ff48_write(uint_fast8_t octet);
-uint_fast8_t wd2797_ff48_read(void);
+void wd2797_ff48_write(unsigned int octet);
+unsigned int wd2797_ff48_read(void);
 
 #endif  /* __WD2797_H__ */
